Added confirm popup for deletebutton for all tabs + changes to warning in btr

+++ fsk-online-ts/src/main/ts/app/tabs/TreatmentWillTab.ts
@@ -82,8 +82,9 @@ export default class TreatmentWillTab extends TemplateWidget implements TabbedPa
             $: isChecked
         };
     }
+
     public async setupBindings(): Promise<void> {
-        this.warningIfLivingWillExist((await this.livingWillCache.loadHasRegistration() && this.isAdministratorUser));
+        this.warningIfLivingWillExist(this.livingWillCache.loadHasRegistration(), this.isAdministratorUser);
 
         this.terminallyIllCheckbox = new CheckboxWrapper(this.getElementByVarName(`terminally-ill-checkbox`));
         this.terminallyIllPanel = this.container.resolve<TreatmentWillWishPanel>(TreatmentWillWishPanel);
@@ -139,9 +140,23 @@ export default class TreatmentWillTab extends TemplateWidget implements TabbedPa
 
         this.deleteButton = new SDSButton("Slet registrering", "danger", async () => {
             try {
-                await this.fskService.deleteTreatmentWillForPatient(this.moduleContext.getPatient().getCpr());
-                this.updateCache(false);
-                this.moduleContext.setApplicationContextId(`PATIENT`);
+                const yesOption = <DialogOption>{
+                    buttonStyle: ButtonStyle.GREEN,
+                    text: `Slet`,
+                };
+
+                const noOption = <DialogOption>{
+                    buttonStyle: ButtonStyle.RED,
+                    text: `Fortryd`,
+                };
+                const yesIsClicked = await PopupDialog.display(PopupDialogKind.WARNING, "Bekræft sletning",
+                    "<p>Er du sikker på du vil slette patientens behandlingstestamenteregistrering?</p>",
+                    noOption, yesOption);
+                if (yesIsClicked == yesOption) {
+                    await this.fskService.deleteTreatmentWillForPatient(this.moduleContext.getPatient().getCpr());
+                    this.updateCache(false);
+                    this.moduleContext.setApplicationContextId(`PATIENT`);
+                }
             } catch (error) {
                 ErrorDisplay.showError("Det skete en fejl", ErrorUtil.getMessage(error));
             }
@@ -158,8 +173,8 @@ export default class TreatmentWillTab extends TemplateWidget implements TabbedPa
         this.rootElement.appendChild(this.element);
     }
 
-    private warningIfLivingWillExist(livingWillExist: boolean) {
-        Widget.setVisible(this.getElementByVarName(`living-will-exists`), livingWillExist);
+    private async warningIfLivingWillExist(livingWillExist: Promise<boolean>, isAdmin: boolean) {
+        Widget.setVisible(this.getElementByVarName(`living-will-exists`), await livingWillExist && isAdmin);
     }
 
     public hideButtons() {
@@ -208,7 +223,7 @@ export default class TreatmentWillTab extends TemplateWidget implements TabbedPa
     }
 
     public async setVisible(visible: boolean): Promise<void> {
-        if(!this.moduleContext.getPatient()) {
+        if (!this.moduleContext.getPatient()) {
             return;
         }
 
@@ -239,9 +254,6 @@ export default class TreatmentWillTab extends TemplateWidget implements TabbedPa
                 yesOption);
             canSee = yesClicked === yesOption;
             this.canSee = canSee;
-        }
-        if (! await this.livingWillCache.loadHasRegistration()) {
-            this.warningIfLivingWillExist(false);
         }
 
         super.setVisible(canSee);
@@ -279,7 +291,7 @@ export default class TreatmentWillTab extends TemplateWidget implements TabbedPa
         const treatmentWillExistsForHealthcareProvider = !treatmentWillExist && this.hasAuthAndNotAdmin;
         if (applicationContextId === "PATIENT" && treatmentWillExistsForHealthcareProvider) {
             this.moduleContext.hideTab(this.ID);
-        } else if(applicationContextId === "PATIENT") {
+        } else if (applicationContextId === "PATIENT") {
             this.moduleContext.showTab(this.ID);
         } else {
             this.moduleContext.hideTab(this.ID);
@@ -300,7 +312,8 @@ export default class TreatmentWillTab extends TemplateWidget implements TabbedPa
         this.deleteButton.setVisible(!isCreateMode && this.isAdministratorUser);
     }
 
-    public setData(treatmentWill: FSKTypes.TreatmentWillType): void {
+    public async setData(treatmentWill: FSKTypes.TreatmentWillType): Promise<void> {
+        this.warningIfLivingWillExist(this.livingWillCache.loadHasRegistration(), true);
         if (treatmentWill) {
             Object.entries(treatmentWill).forEach(([property, will]) => {
                 switch (property) {
@@ -356,7 +369,6 @@ export default class TreatmentWillTab extends TemplateWidget implements TabbedPa
             ? null
             : wishPanel.getValue();
     }
-
 
     private addListeners() {
         if (!this.treatmentWillChangeHandler) {
